spec: inline & drop py_package_prefix
This variable is hardcoded as "python3" on SUSE distros, and on others
(EL and Fedora) it is set using an helper variable. In the end the
result is always "python3", and there were bits already hardcoding that
instead of using "py_package_prefix".

Since the hardcoded "python3" prefix has been working fine so far,
simply put "python3" everywhere instead of the "py_package_prefix"
variable, and drop it altogether.